refactor(editor): Extract NodeSettings code into features/ndv (no-changelog)
