+++ packages/frontend/editor-ui/src/features/ndv/components/settings/NodeSettingsTabs.vue
